Simplify instrument signature
